+++ api/source/pastemyst/auth/github.d
@@ -133,6 +133,42 @@ public class AuthGitHubAPI : IAuthGitHubAPI
     }
 }
 
+/++
+ + GitHub API interface.
+ +/
+public interface IGitHubAPI
+{
+    /++
+     + GET /user
+     +
+     + Returns the current authenticated user.
+     +/
+    @path ("/user")
+    @headerParam ("authorization", "Authorization")
+    Json getUser (string authorization) @safe;
+}
+
+/++
+ + GitHub API interface.
+ +/
+public class GitHubAPI : IGitHubAPI
+{
+    /++
+     + GET /user
+     +
+     + Returns the current authenticated user.
+     +/
+    public Json getUser (string authorization) @safe
+    {
+        import vibe.data.json : serializeToJson;
+
+        // TODO: Handle if the wrong format for authorization is provided
+        string token = authorization ["Bearer ".length..$];
+
+        return getGitHubUserJwt (token).serializeToJson ();
+    }
+}
+
 private User getGitHubUser (string accessToken) @safe
 {
     User user;
@@ -148,6 +184,7 @@ private User getGitHubUser (string accessToken) @safe
 
         user.id = json ["id"].get!int;
         user.username = json ["login"].get!string;
+        user.avatar = json ["avatar_url"].get!string;
     });
 
     return user;
