+++ api/source/pastemyst/data/user.d
@@ -17,4 +17,9 @@ public struct User
      + User's username
      +/
     public string username;
+
+    /++
+     + User's avatar url
+     +/
+    public string avatar;
 }
