+++ frontend/src/scripts/api/auth.ts
@@ -1,12 +1,13 @@
 import { apiEndpoint } from "api";
+import { getJwt } from "security/cookies";
 
 export async function isLoggedIn (): Promise<boolean>
 {
-    const jwtCookie: string = getCookie ("github");
+    const jwtCookie: string = getJwt ();
 
-    if (jwtCookie === null)
+    if (jwtCookie === "")
     {
-        console.log ("reee");
+        return false;
     }
     else
     {
@@ -21,37 +22,4 @@ export async function isLoggedIn (): Promise<boolean>
 
         return JSON.parse (await response.text ());
     }
-}
-
-export async function getJwt (): Promise<string>
-{
-    const jwtCookie: string = getCookie ("github");
-
-    if (jwtCookie !== null)
-    {
-        return jwtCookie;
-    }
-}
-
-function getCookie (name: string): string
-{
-    const nameEQ: string = name + "=";
-    const cookies: string [] = document.cookie.split(";");
-    
-    for (let i: number = 0; i < cookies.length; i++)
-    {
-        let cookie: string = cookies [i];
-        
-        while (cookie.charAt (0) === " ")
-        {
-            cookie = cookie.substring (1, cookie.length);
-        }
-
-        if (cookie.indexOf (nameEQ) === 0)
-        {
-            return cookie.substring (nameEQ.length, cookie.length);
-        }
-    }
-
-    return null;
 }
