+++ frontend/src/scripts/data/user.ts
@@ -1,0 +1,6 @@
+export default class User
+{
+    public id: number;
+    public username: string;
+    public avatar: string;
+}
