+++ frontend/src/styles/styles.scss
@@ -15,6 +15,7 @@
 
 @import "pages/home";
 @import "pages/paste";
+@import "pages/profile";
 
 body
 {
