+++ frontend/src/views/profile.pug
@@ -2,6 +2,9 @@ include components/header
 
 include components/navigation
 
-h1 profile
+#profile-header
+    img.avatar
+    p.username
+    a.logout logout
 
 include components/footer
